+++ CHANGELOG.md
@@ -5,7 +5,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 [Semantic Versioning]. For IRMA vs IRMA-core compatibility, please see the
 [version matrix](VERSION_MATRIX.md).
 
-## [0.6.1] - TBD
+## [0.6.1] - 2025-09-09
 
 ### Changed
 
@@ -14,8 +14,9 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 
 ### Fixes
 
-- SRA paired-end headers could be could be incorrectly flagged as unmated when
-  `--filter-widows` was selected in `trimmer` or `sampler`. This has been fixed.
+- Fixes bug where SRA paired-end headers could be could be incorrectly flagged
+  as unmated when `--filter-widows` was selected in `trimmer`, or with paired
+  reads in `sampler`
 - Fixes a reproducibility issue in `sampler` where the seed wasn't being used in
   one case